- added actual accuracy calc in AnnHarness. Updated it to test on the mnist test set. - Moved all the MnistReader classes to their own files. - simplified SoftMax calculation. - All error from the ILossFunction are now scalars not vectors. Turns out the code always summed them up anyway or took the mean. so now they're all scalars. - deleted VanillaCrossEntropy. - added some tests - all biases are now initialized with -0.1f to 0.1f - added get test pairs from MnistTrainingSet

+++ MatrixLibrary/Base Classes/ColumnVectorBase.cs
@@ -21,7 +21,7 @@ namespace MatrixLibrary.BaseClasses
             }
             return max;
         }
-        public void SetRandom(int seed, int min, int max)
+        public void SetRandom(int seed, float min, float max)
         {
             Random rnd = new Random(seed);
             float width = max - min;
